+++ lib/src/ext/extensions.dart
@@ -36,6 +36,7 @@ import 'package:gltf/src/ext/KHR_materials_volume/khr_materials_volume.dart';
 import 'package:gltf/src/ext/KHR_mesh_quantization/khr_mesh_quantization.dart';
 import 'package:gltf/src/ext/KHR_texture_transform/khr_texture_transform.dart';
 import 'package:gltf/src/ext/VRMC_materials_mtoon/vrmc_materials_mtoon.dart';
+import 'package:gltf/src/ext/VRMC_node_constraint/vrmc_node_constraint.dart';
 import 'package:gltf/src/ext/VRMC_springBone/vrmc_spring_bone.dart';
 import 'package:gltf/src/ext/VRMC_vrm/vrmc_vrm.dart';
 import 'package:gltf/src/hash.dart';
@@ -130,6 +131,7 @@ const List<Extension> kDefaultExtensions = <Extension>[
   khrMeshQuantizationExtension,
   khrTextureTransformExtension,
   vrmcMaterialMtoon,
+  vrmcNodeConstraint,
   vrmcSpringBone,
   vrmcVrm,
 ];